+++ lib/hitimes/mutexed_stats.rb
@@ -11,22 +11,7 @@ module Hitimes
   # Ruby the Stats object is already threadsafe, so there is no need to use
   # MutexedStats.
   #
-  class MutexedStats < Stats
-    def initialize
-      @mutex = Mutex.new
-    end
-
-    # call-seq:
-    #   mutex_stat.update( val ) -> nil
-    # 
-    # Update the running stats with the new value in a threadsafe manner.
-    #
-    def update( value )
-      @mutex.synchronize do
-        super( value )
-      end
-    end
-  end
+  MutexedStats = Stats
 end