+++ spec/mutex_stats_spec.rb
@@ -16,21 +16,14 @@ describe Hitimes::MutexedStats do
     return stats
   end
 
-  if (not defined? RUBY_ENGINE) or (RUBY_ENGINE == "ruby") then
-    it "Hitimes::Stats is threadsafe" do
-      stats = run_with_scissors( ::Hitimes::Stats.new, @threads, @iters )
-      stats.count.must_equal @final_value
-    end
-  else
-    it "Hitimes::Stats is not threadsafe" do
-      stats = run_with_scissors( ::Hitimes::Stats.new, @threads, @iters )
-      stats.count.wont_equal @final_value
-    end
+  it "Hitimes::Stats is threadsafe" do
+    stats = run_with_scissors( ::Hitimes::Stats.new, @threads, @iters )
+    _(stats.count).must_equal @final_value
   end
 
   it "has a threadsafe update" do
     stats = run_with_scissors( ::Hitimes::MutexedStats.new, @threads, @iters )
-    stats.count.must_equal @final_value
+    _(stats.count).must_equal @final_value
   end
 
 end